requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
openai
langchain-ollama
requests
langchain-core